+++ src/sphinx_codeautolink/__init__.py
@@ -3,10 +3,7 @@ import os as _os
 from pathlib import Path as _Path
 
 from sphinx.application import Sphinx
-from .extension import SphinxCodeAutoLink
-from .extension.directive import (
-    CodeReferences, ImplicitImport, ConcatBlocks, AutoLinkSkip
-)
+from .extension import backref, directive, SphinxCodeAutoLink
 
 _version_file = _Path(_os.path.realpath(__file__)).parent / "VERSION"
 __version__ = _version_file.read_text().strip()
@@ -18,14 +15,21 @@ def setup(app: Sphinx):
     app.add_css_file('sphinx-codeautolink.css')
     app.add_config_value('codeautolink_autodoc_inject', True, 'html', types=[bool])
 
-    app.add_directive('concat-blocks', ConcatBlocks)
-    app.add_directive('code-refs', CodeReferences)
-    app.add_directive('implicit-import', ImplicitImport)
-    app.add_directive('autolink-skip', AutoLinkSkip)
+    app.add_directive('concat-blocks', directive.ConcatBlocks)
+    app.add_directive('code-refs', directive.CodeReferences)
+    app.add_directive('implicit-import', directive.ImplicitImport)
+    app.add_directive('autolink-skip', directive.AutoLinkSkip)
 
     app.connect('builder-inited', state.build_inited)
     app.connect('autodoc-process-docstring', state.autodoc_process_docstring)
     app.connect('doctree-read', state.parse_blocks)
     app.connect('doctree-resolved', state.generate_backref_tables)
     app.connect('build-finished', state.apply_links)
+
+    app.add_node(
+        backref.DetailsNode, html=(backref.visit_details, backref.depart_details)
+    )
+    app.add_node(
+        backref.SummaryNode, html=(backref.visit_summary, backref.depart_summary)
+    )
     return {'version': __version__}